# print_corners