In this PR, the recursive method used in `dpctl.tensor._print._nd_corners` function is replaced with an iterative method to improve performance.
`x_dpt = dpt.reshape(dpt.arange(6*6*117*117, dtype='i4'),(6,117,117,6))`
`%timeit -r 20 dpt.usm_ndarray_repr(x_dpt)`

New timing: 4.55 ms ± 645 µs per loop (mean ± std. dev. of 20 runs, 100 loops each)
Old timing: 6.43 ms ± 2.31 ms per loop (mean ± std. dev. of 20 runs, 100 loops each)

- [x] Have you provided a meaningful PR description?
- [x] Have you added a test, reproducer or referred to an issue with a reproducer?
- [x] Have you tested your changes locally for CPU and GPU devices?
- [x] Have you made sure that new changes do not introduce compiler warnings?
- [x] Have you checked performance impact of proposed changes?
- [ ] If this PR is a work in progress, are you opening the PR as a draft?