@oleksandr-pavlyk 2023-04-25
For array of rank `r` there are $2^r$ corners instead of $r^2$.

But in our case the `r` need not be the whole rank, we should only consider dimensions which size exceeds the threshold.

Also, instead of decoding integer to sequence of bits using a string, perhaps consider using `itertools.outer`.